Optimize Push and Pop
We don't need to check if queue is full every byte since we are forcing
the caller to size their buffer to be an integer multiple of the dataSize.
We also don't need to check for the 'stash' condition every byte for
the same reason. This optimization increases execution time by
roughly 33%.